LiveFeed context added. Need to sort out the archives and camera status history
